MLServer/requirements.txt
absl-py==0.2.0
astor==0.6.2
backports.weakref==1.0.post1
bleach==1.5.0
certifi==2018.4.16
chardet==3.0.4
Django==1.11.12
django-allauth==0.35.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.11.0
html5lib==0.9999999
idna==2.6
Markdown==2.6.11
mock==2.0.0
mysql-connector-python==8.0.6
mysqlclient==1.3.12
numpy==1.14.2
oauthlib==2.0.7
pbr==4.0.2
pkg-resources==0.0.0
protobuf==3.5.2.post1
PyJWT==1.6.1
python-openid==2.2.5
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
